DogGo: Authentication

+++ DogGo/Controllers/DogsController.cs
@@ -1,9 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using DogGo.Models;
 using DogGo.Repositories;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
@@ -23,9 +25,12 @@ namespace DogGo.Controllers
         }
 
         // GET: Dogs
+        [Authorize]
         public ActionResult Index()
         {
-            List<Dog> dogs = _dogRepo.GetAllDogs();
+            int ownerId = GetCurrentUserId();
+
+            List<Dog> dogs = _dogRepo.GetAllDogsByOwner(ownerId);
 
             return View(dogs);
         }
@@ -45,6 +50,7 @@ namespace DogGo.Controllers
         }
 
         // GET: Dogs/Create
+        [Authorize]
         public ActionResult Create()
         {
             return View();
@@ -53,27 +59,32 @@ namespace DogGo.Controllers
         // POST: Dogs/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create(Dog Dog)
+        [Authorize]
+        public ActionResult Create(Dog dog)
         {
             try
             {
-                _dogRepo.AddDog(Dog);
+                // update the dogs OwnerId to the current user's Id 
+                dog.OwnerId = GetCurrentUserId();
+
+                _dogRepo.AddDog(dog);
 
                 return RedirectToAction("Index");
             }
             catch (Exception ex)
             {
-                return View(Dog);
+                return View(dog);
             }
         }
 
         // GET: Dogs/Edit/5
+        [Authorize]
         public ActionResult Edit(int id)
         {
             Dog dog = _dogRepo.GetDogById(id);
             dog.Owner = _ownerRepo.GetOwnerById(dog.OwnerId);
 
-            if (dog == null)
+            if (dog == null || dog.OwnerId != GetCurrentUserId())
             {
                 return NotFound();
             }
@@ -84,13 +95,21 @@ namespace DogGo.Controllers
         // POST: Dogs/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize]
         public ActionResult Edit(int id, Dog dog)
         {
             try
             {
-                _dogRepo.UpdateDog(dog);
+                if (dog.OwnerId == GetCurrentUserId())
+                {
+                    _dogRepo.UpdateDog(dog);
 
-                return RedirectToAction("Index");
+                    return RedirectToAction("Index");
+                }
+                else
+                {
+                    return NotFound();
+                }
             }
             catch (Exception ex)
             {
@@ -99,29 +118,54 @@ namespace DogGo.Controllers
         }
 
         // GET: Dogs/Delete/5
+        [Authorize]
         public ActionResult Delete(int id)
         {
             Dog dog = _dogRepo.GetDogById(id);
             dog.Owner = _ownerRepo.GetOwnerById(dog.OwnerId);
 
-            return View(dog);
+            if (dog.OwnerId == GetCurrentUserId())
+            {
+
+                return View(dog);
+            }
+            else
+            {
+                return NotFound();
+            }
         }
 
         // POST: Dogs/Delete/5
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize]
         public ActionResult Delete(int id, Dog dog)
         {
+            dog = _dogRepo.GetDogById(id);
+            
             try
             {
-                _dogRepo.DeleteDog(id);
+                if (dog.OwnerId == GetCurrentUserId())
+                {
+                    _dogRepo.DeleteDog(id);
 
-                return RedirectToAction("Index");
+                    return RedirectToAction("Index");
+                }
+                else
+                {
+                    return NotFound();
+                }
             }
             catch (Exception ex)
             {
                 return View(dog);
             }
+        }
+
+        private int GetCurrentUserId()
+        {
+            string id = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            return int.Parse(id);
         }
     }
 }
